+++ src/Command/Game/Mode/AbstractMode.php
@@ -3,10 +3,10 @@
 namespace ChessServer\Command\Game\Mode;
 
 use ChessServer\Command\Game\Game;
-use ChessServer\Command\Game\Async\PlayCommand;
-use ChessServer\Command\Game\Async\PlayLanCommand;
-use ChessServer\Command\Game\Async\LegalCommand;
-use ChessServer\Command\Game\Async\UndoCommand;
+use ChessServer\Command\Game\NonBlocking\PlayCommand;
+use ChessServer\Command\Game\NonBlocking\PlayLanCommand;
+use ChessServer\Command\Game\NonBlocking\LegalCommand;
+use ChessServer\Command\Game\NonBlocking\UndoCommand;
 use Firebase\JWT\JWT;
 use Firebase\JWT\Key;
