+++ src/Command/Game/NonBlocking/RematchCommand.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ChessServer\Command\Game\Async
+namespace ChessServer\Command\Game\NonBlocking
 ;
 
 use ChessServer\Command\AbstractNonBlockingCommand;
